refactor(cpu): drop CLPEAK_CPU_NATIVE_ARCH, modularize kernels, keep Apple M2+ dtypes
Remove the native-arch build mode entirely -- the CPU backend now has a single
portable path (per-ISA feature TUs + runtime dispatch), so one binary is safe on
any CPU and still uses the best available ISA.

- split the 966-line cpu_kernels_impl.h into src/cpu/kernels/{base,lowp,matrix,
  sve}_compute.h; it is now just the per-TU tuTable() aggregator
- add cpu_tu_registry.h (CLPEAK_TU_REGISTRY X-macro) as the single tag list that
  drives the dispatcher's unconditional table-accessor forward declarations
- gate Windows-SVE behind one CLPEAK_CPU_ENABLE_WIN_SVE toggle (_clpeak_sve_ok)
- add an Apple arm64 bf16/i8mm TU pair so M2+ keeps bf16/BFMMLA/SMMLA rows via
  the portable path (runtime-gated by sysctlbyname; stays Unsupported on M1)
- drop the dead `neon` TU wiring
