+++ src/cpu/cpu_kernels_impl.h
@@ -2,894 +2,41 @@
 #define CPU_KERNELS_IMPL_H
 
 // ===========================================================================
-// All compute / read kernel bodies, plus a per-TU table builder.  This header
-// is compiled once per feature TU (cpu_kernels_tu.cpp, built N times with
-// different -m/-arch flags); cpu_simd.h selects the SIMD path from those flags,
-// and the #if-gated advanced kernels compile in only when the TU's flags enable
-// the feature.  Everything lives in an anonymous namespace so each TU gets its
-// own internal copy (no ODR clash, no cross-TU inlining).
+// Per-TU kernel aggregation.  This header is compiled once per feature TU
+// (cpu_kernels_tu.cpp, built N times with different -m/-arch flags); it pulls in
+// the kernel bodies -- grouped into cohesive headers under kernels/ -- and emits
+// this TU's CpuKernelTable from whatever kernels its build flags enabled.
+//
+//   kernels/base_compute.h    fp32 / fp64 / int32 chains + streaming read (all TUs)
+//   kernels/lowp_compute.h    fp16 / bf16-dot / mixed-precision / int8-dot / bf16-FMA
+//   kernels/matrix_compute.h  AMX (int8/bf16/fp16/tf32/fp8) + NEON SMMLA/BFMMLA
+//   kernels/sve_compute.h     ARM SVE compute + SVE bf16/i8mm matrix
+//
+// cpu_simd.h selects the base SIMD path from the TU's flags, and each advanced
+// kernel is #if-gated on the compile-feature macros those flags define, so a
+// given TU compiles in only the kernels it can run.  Everything lives in an
+// anonymous namespace so each TU gets its own internal copy (no ODR clash, no
+// cross-TU inlining).  To add a new TU: give it a kernel body in the matching
+// kernels/ header (a new #if block + CPU_HAS_<X>/CPU_MAT_<X> define) and a table
+// slot below; then register the TU in CMakeLists.txt + cpu_dispatch.cpp.
 // ===========================================================================
 
 #include "cpu_kernels.h"
 #include "cpu_simd.h"
 
-#include <cstring>
+#include "kernels/base_compute.h"
+#include "kernels/lowp_compute.h"
+#include "kernels/matrix_compute.h"
+#include "kernels/sve_compute.h"
 
 #ifndef CLPEAK_ISA_NAME_STR
 #define CLPEAK_ISA_NAME_STR "scalar"
-#endif
-
-#if (defined(__AMX_INT8__) || defined(__AMX_BF16__)) && (defined(__x86_64__) || defined(_M_X64))
-#include <immintrin.h>
-#endif
-
-#if defined(__ARM_FEATURE_SVE)
-#include <arm_sve.h>
 #endif
 
 namespace clpeak_cpu {
 namespace {
 
 using namespace cpu_simd;
-
-// ---- FP32 / FP64 FMA-chains -----------------------------------------------
-static double runFp32Chain(uint64_t outer)
-{
-  f32v acc[F32_NACC];
-  // Seed the coefficients through volatile so the compiler can't prove their
-  // values and close the affine recurrence acc=acc*b+c.  On non-FMA targets
-  // (SSE2 / scalar) f32_fma is a transparent mul+add; with -ffast-math the
-  // CPU_UNROLL_K-unrolled chain would otherwise fold N steps into one
-  // (acc*b^N + const), deleting most of the work while opsPerIter still counts
-  // it -> an impossible peak (SSE2 reported > AVX2).  Same guard as the int32
-  // chain.  No-op on FMA targets (b/c become runtime operands loaded once).
-  volatile float vb = 0.999999f, vc = 0.000001f;
-  const f32v b = f32_set(vb);
-  const f32v c = f32_set(vc);
-  for (int j = 0; j < F32_NACC; j++) acc[j] = f32_set(0.1f * (float)(j + 1));
-  for (uint64_t o = 0; o < outer; o++)
-    CPU_UNROLL_K
-    for (int k = 0; k < INNER; k++)
-    {
-      CPU_UNROLL_FULL
-      for (int j = 0; j < F32_NACC; j++) acc[j] = f32_fma(acc[j], b, c);
-    }
-  f32v s = acc[0];
-  for (int j = 1; j < F32_NACC; j++) s = f32_add(s, acc[j]);
-  return (double)f32_hsum(s);
-}
-
-static double runFp64Chain(uint64_t outer)
-{
-  f64v acc[F64_NACC];
-  // Volatile-seed the coefficients to keep the affine chain from collapsing on
-  // non-FMA (SSE2 / scalar) targets under -ffast-math.  See runFp32Chain.
-  volatile double vb = 0.999999, vc = 0.000001;
-  const f64v b = f64_set(vb);
-  const f64v c = f64_set(vc);
-  for (int j = 0; j < F64_NACC; j++) acc[j] = f64_set(0.1 * (double)(j + 1));
-  for (uint64_t o = 0; o < outer; o++)
-    CPU_UNROLL_K
-    for (int k = 0; k < INNER; k++)
-    {
-      CPU_UNROLL_FULL
-      for (int j = 0; j < F64_NACC; j++) acc[j] = f64_fma(acc[j], b, c);
-    }
-  f64v s = acc[0];
-  for (int j = 1; j < F64_NACC; j++) s = f64_add(s, acc[j]);
-  return (double)f64_hsum(s);
-}
-
-// ---- INT32 multiply-accumulate chain --------------------------------------
-static double runInt32Chain(uint64_t outer)
-{
-  i32v acc[I32_NACC];
-  // Opaque multiplier: a compile-time-constant multiplier lets the compiler
-  // strength-reduce `acc*k` into shifts/adds (measuring the shifter, not the
-  // integer multiplier, and varying with -mtune/inlining).  Reading it through
-  // volatile forces a real vpmulld / vmlaq — a consistent, honest IMAD peak.
-  volatile int vmul = 1664525;
-  const i32v b = i32_set((int)vmul);
-  const i32v c = i32_set(1013904223);
-  for (int j = 0; j < I32_NACC; j++) acc[j] = i32_set(j + 1);
-  for (uint64_t o = 0; o < outer; o++)
-    CPU_UNROLL_K
-    for (int k = 0; k < INNER; k++)
-    {
-      CPU_UNROLL_FULL
-      for (int j = 0; j < I32_NACC; j++) acc[j] = i32_madd(acc[j], b, c);
-    }
-  i32v s = acc[0];
-  for (int j = 1; j < I32_NACC; j++) s = i32_add(s, acc[j]);
-  return (double)i32_hsum(s);
-}
-
-// ---- Streaming read (integer XOR checksum) --------------------------------
-static uint64_t readBufferChecksum(const float *p, size_t M, uint64_t iters)
-{
-#if defined(__AVX512F__)
-  constexpr size_t W = 16;
-  const size_t step = 8 * W;
-  __m512i x0 = _mm512_setzero_si512(), x1 = x0, x2 = x0, x3 = x0;
-  __m512i x4 = x0, x5 = x0, x6 = x0, x7 = x0;
-  uint64_t tail = 0;
-  for (uint64_t it = 0; it < iters; it++)
-  {
-    size_t i = 0;
-    for (; i + step <= M; i += step)
-    {
-      x0 = _mm512_xor_si512(x0, _mm512_castps_si512(_mm512_loadu_ps(p + i + 0 * W)));
-      x1 = _mm512_xor_si512(x1, _mm512_castps_si512(_mm512_loadu_ps(p + i + 1 * W)));
-      x2 = _mm512_xor_si512(x2, _mm512_castps_si512(_mm512_loadu_ps(p + i + 2 * W)));
-      x3 = _mm512_xor_si512(x3, _mm512_castps_si512(_mm512_loadu_ps(p + i + 3 * W)));
-      x4 = _mm512_xor_si512(x4, _mm512_castps_si512(_mm512_loadu_ps(p + i + 4 * W)));
-      x5 = _mm512_xor_si512(x5, _mm512_castps_si512(_mm512_loadu_ps(p + i + 5 * W)));
-      x6 = _mm512_xor_si512(x6, _mm512_castps_si512(_mm512_loadu_ps(p + i + 6 * W)));
-      x7 = _mm512_xor_si512(x7, _mm512_castps_si512(_mm512_loadu_ps(p + i + 7 * W)));
-    }
-    for (; i < M; i++) { uint32_t v; std::memcpy(&v, p + i, sizeof(v)); tail ^= v; }
-  }
-  __m512i x = _mm512_xor_si512(_mm512_xor_si512(_mm512_xor_si512(x0, x1), _mm512_xor_si512(x2, x3)),
-                               _mm512_xor_si512(_mm512_xor_si512(x4, x5), _mm512_xor_si512(x6, x7)));
-  alignas(64) uint64_t tmp[8]; _mm512_store_si512((__m512i *)tmp, x);
-  for (uint64_t v : tmp) tail ^= v;
-  return tail;
-#elif defined(__AVX2__)
-  constexpr size_t W = 8;
-  const size_t step = 8 * W;
-  __m256i x0 = _mm256_setzero_si256(), x1 = x0, x2 = x0, x3 = x0;
-  __m256i x4 = x0, x5 = x0, x6 = x0, x7 = x0;
-  uint64_t tail = 0;
-  for (uint64_t it = 0; it < iters; it++)
-  {
-    size_t i = 0;
-    for (; i + step <= M; i += step)
-    {
-      x0 = _mm256_xor_si256(x0, _mm256_castps_si256(_mm256_loadu_ps(p + i + 0 * W)));
-      x1 = _mm256_xor_si256(x1, _mm256_castps_si256(_mm256_loadu_ps(p + i + 1 * W)));
-      x2 = _mm256_xor_si256(x2, _mm256_castps_si256(_mm256_loadu_ps(p + i + 2 * W)));
-      x3 = _mm256_xor_si256(x3, _mm256_castps_si256(_mm256_loadu_ps(p + i + 3 * W)));
-      x4 = _mm256_xor_si256(x4, _mm256_castps_si256(_mm256_loadu_ps(p + i + 4 * W)));
-      x5 = _mm256_xor_si256(x5, _mm256_castps_si256(_mm256_loadu_ps(p + i + 5 * W)));
-      x6 = _mm256_xor_si256(x6, _mm256_castps_si256(_mm256_loadu_ps(p + i + 6 * W)));
-      x7 = _mm256_xor_si256(x7, _mm256_castps_si256(_mm256_loadu_ps(p + i + 7 * W)));
-    }
-    for (; i < M; i++) { uint32_t v; std::memcpy(&v, p + i, sizeof(v)); tail ^= v; }
-  }
-  __m256i x = _mm256_xor_si256(_mm256_xor_si256(_mm256_xor_si256(x0, x1), _mm256_xor_si256(x2, x3)),
-                               _mm256_xor_si256(_mm256_xor_si256(x4, x5), _mm256_xor_si256(x6, x7)));
-  alignas(32) uint64_t tmp[4]; _mm256_store_si256((__m256i *)tmp, x);
-  for (uint64_t v : tmp) tail ^= v;
-  return tail;
-#elif defined(__SSE2__)
-  constexpr size_t W = 4;
-  const size_t step = 8 * W;
-  __m128i x0 = _mm_setzero_si128(), x1 = x0, x2 = x0, x3 = x0;
-  __m128i x4 = x0, x5 = x0, x6 = x0, x7 = x0;
-  uint64_t tail = 0;
-  for (uint64_t it = 0; it < iters; it++)
-  {
-    size_t i = 0;
-    for (; i + step <= M; i += step)
-    {
-      x0 = _mm_xor_si128(x0, _mm_castps_si128(_mm_loadu_ps(p + i + 0 * W)));
-      x1 = _mm_xor_si128(x1, _mm_castps_si128(_mm_loadu_ps(p + i + 1 * W)));
-      x2 = _mm_xor_si128(x2, _mm_castps_si128(_mm_loadu_ps(p + i + 2 * W)));
-      x3 = _mm_xor_si128(x3, _mm_castps_si128(_mm_loadu_ps(p + i + 3 * W)));
-      x4 = _mm_xor_si128(x4, _mm_castps_si128(_mm_loadu_ps(p + i + 4 * W)));
-      x5 = _mm_xor_si128(x5, _mm_castps_si128(_mm_loadu_ps(p + i + 5 * W)));
-      x6 = _mm_xor_si128(x6, _mm_castps_si128(_mm_loadu_ps(p + i + 6 * W)));
-      x7 = _mm_xor_si128(x7, _mm_castps_si128(_mm_loadu_ps(p + i + 7 * W)));
-    }
-    for (; i < M; i++) { uint32_t v; std::memcpy(&v, p + i, sizeof(v)); tail ^= v; }
-  }
-  __m128i x = _mm_xor_si128(_mm_xor_si128(_mm_xor_si128(x0, x1), _mm_xor_si128(x2, x3)),
-                            _mm_xor_si128(_mm_xor_si128(x4, x5), _mm_xor_si128(x6, x7)));
-  alignas(16) uint32_t tmp[4]; _mm_store_si128((__m128i *)tmp, x);
-  for (uint32_t v : tmp) tail ^= v;
-  return tail;
-#elif defined(__aarch64__) || defined(_M_ARM64)
-  // AArch64 only (matches cpu_simd.h, incl. the MSVC ARM64 _M_ARM64 alias):
-  // 32-bit ARMv7 doesn't pull in arm_neon.h, so it uses the scalar read
-  // fallback below.
-  constexpr size_t W = 4;
-  const size_t step = 8 * W;
-  uint32x4_t x0 = vdupq_n_u32(0), x1 = x0, x2 = x0, x3 = x0;
-  uint32x4_t x4 = x0, x5 = x0, x6 = x0, x7 = x0;
-  uint64_t tail = 0;
-  for (uint64_t it = 0; it < iters; it++)
-  {
-    size_t i = 0;
-    for (; i + step <= M; i += step)
-    {
-      x0 = veorq_u32(x0, vreinterpretq_u32_f32(vld1q_f32(p + i + 0 * W)));
-      x1 = veorq_u32(x1, vreinterpretq_u32_f32(vld1q_f32(p + i + 1 * W)));
-      x2 = veorq_u32(x2, vreinterpretq_u32_f32(vld1q_f32(p + i + 2 * W)));
-      x3 = veorq_u32(x3, vreinterpretq_u32_f32(vld1q_f32(p + i + 3 * W)));
-      x4 = veorq_u32(x4, vreinterpretq_u32_f32(vld1q_f32(p + i + 4 * W)));
-      x5 = veorq_u32(x5, vreinterpretq_u32_f32(vld1q_f32(p + i + 5 * W)));
-      x6 = veorq_u32(x6, vreinterpretq_u32_f32(vld1q_f32(p + i + 6 * W)));
-      x7 = veorq_u32(x7, vreinterpretq_u32_f32(vld1q_f32(p + i + 7 * W)));
-    }
-    for (; i < M; i++) { uint32_t v; std::memcpy(&v, p + i, sizeof(v)); tail ^= v; }
-  }
-  uint32x4_t x = veorq_u32(veorq_u32(veorq_u32(x0, x1), veorq_u32(x2, x3)),
-                           veorq_u32(veorq_u32(x4, x5), veorq_u32(x6, x7)));
-  alignas(16) uint32_t tmp[4]; vst1q_u32(tmp, x);
-  for (uint32_t v : tmp) tail ^= v;
-  return tail;
-#else
-  uint64_t acc = 0;
-  for (uint64_t it = 0; it < iters; it++)
-    for (size_t i = 0; i < M; i++) { uint32_t v; std::memcpy(&v, p + i, sizeof(v)); acc ^= v; }
-  return acc;
-#endif
-}
-
-// Advanced-dtype kernels.  Each is gated on a specific compile feature, so a
-// dedicated feature TU enables exactly one.  CLPEAK_CORE_ONLY (passed to MSVC
-// TUs, whose coarse /arch:AVX512 can't isolate sub-features) keeps these out so
-// a core-AVX-512 binary never emits VNNI/BF16 it might lack at runtime.
-#ifndef CLPEAK_CORE_ONLY
-
-// ---- FP16 native FMA -------------------------------------------------------
-#if defined(__ARM_FEATURE_FP16_VECTOR_ARITHMETIC)
-#define CPU_HAS_FP16_KERNEL 1
-static constexpr int FP16_LANES = 8, FP16_NACC = 16;
-static double runFp16Chain(uint64_t outer)
-{
-  float16x8_t acc[FP16_NACC];
-  const float16x8_t b = vdupq_n_f16((float16_t)0.9995f);
-  const float16x8_t c = vdupq_n_f16((float16_t)0.001f);
-  for (int j = 0; j < FP16_NACC; j++) acc[j] = vdupq_n_f16((float16_t)(0.1f * (j + 1)));
-  for (uint64_t o = 0; o < outer; o++)
-    CPU_UNROLL_K
-    for (int k = 0; k < INNER; k++)
-    {
-      CPU_UNROLL_FULL
-      for (int j = 0; j < FP16_NACC; j++) acc[j] = vfmaq_f16(c, acc[j], b);
-    }
-  float16x8_t s = acc[0];
-  for (int j = 1; j < FP16_NACC; j++) s = vaddq_f16(s, acc[j]);
-  float16_t tmp[8]; vst1q_f16(tmp, s);
-  double r = 0.0; for (int k = 0; k < 8; k++) r += (double)tmp[k];
-  return r;
-}
-#elif defined(__AVX512FP16__)
-#define CPU_HAS_FP16_KERNEL 1
-static constexpr int FP16_LANES = 32, FP16_NACC = 16;
-static double runFp16Chain(uint64_t outer)
-{
-  __m512h acc[FP16_NACC];
-  const __m512h b = _mm512_set1_ph((_Float16)0.9995f);
-  const __m512h c = _mm512_set1_ph((_Float16)0.001f);
-  for (int j = 0; j < FP16_NACC; j++) acc[j] = _mm512_set1_ph((_Float16)(0.1f * (j + 1)));
-  for (uint64_t o = 0; o < outer; o++)
-    CPU_UNROLL_K
-    for (int k = 0; k < INNER; k++)
-    {
-      CPU_UNROLL_FULL
-      for (int j = 0; j < FP16_NACC; j++) acc[j] = _mm512_fmadd_ph(acc[j], b, c);
-    }
-  __m512h s = acc[0];
-  for (int j = 1; j < FP16_NACC; j++) s = _mm512_add_ph(s, acc[j]);
-  _Float16 tmp[32]; _mm512_storeu_ph(tmp, s);
-  double r = 0.0; for (int k = 0; k < 32; k++) r += (double)tmp[k];
-  return r;
-}
-#endif
-
-// ---- BF16 dot --------------------------------------------------------------
-#if defined(__ARM_FEATURE_BF16_VECTOR_ARITHMETIC) || defined(__ARM_FEATURE_BF16)
-#define CPU_HAS_BF16_KERNEL 1
-static constexpr int BF16_NACC = 16, BF16_FLOPS_PER_INSTR = 16;
-static double runBf16Chain(uint64_t outer)
-{
-  float32x4_t acc[BF16_NACC];
-  const bfloat16x4_t lo = vcvt_bf16_f32(vdupq_n_f32(1.0001f));
-  const bfloat16x8_t a  = vcombine_bf16(lo, lo);
-  const bfloat16x8_t b  = vcombine_bf16(lo, lo);
-  for (int j = 0; j < BF16_NACC; j++) acc[j] = vdupq_n_f32(0.0f);
-  for (uint64_t o = 0; o < outer; o++)
-    CPU_UNROLL_K
-    for (int k = 0; k < INNER; k++)
-    {
-      CPU_UNROLL_FULL
-      for (int j = 0; j < BF16_NACC; j++) acc[j] = vbfdotq_f32(acc[j], a, b);
-    }
-  float32x4_t s = acc[0];
-  for (int j = 1; j < BF16_NACC; j++) s = vaddq_f32(s, acc[j]);
-  return (double)vaddvq_f32(s);
-}
-#elif defined(__AVX512BF16__)
-#define CPU_HAS_BF16_KERNEL 1
-static constexpr int BF16_NACC = 16, BF16_FLOPS_PER_INSTR = 64;
-static double runBf16Chain(uint64_t outer)
-{
-  __m512 acc[BF16_NACC];
-  const __m512bh a = _mm512_cvtne2ps_pbh(_mm512_set1_ps(1.0001f), _mm512_set1_ps(1.0001f));
-  const __m512bh b = a;
-  for (int j = 0; j < BF16_NACC; j++) acc[j] = _mm512_setzero_ps();
-  for (uint64_t o = 0; o < outer; o++)
-    CPU_UNROLL_K
-    for (int k = 0; k < INNER; k++)
-    {
-      CPU_UNROLL_FULL
-      for (int j = 0; j < BF16_NACC; j++) acc[j] = _mm512_dpbf16_ps(acc[j], a, b);
-    }
-  __m512 s = acc[0];
-  for (int j = 1; j < BF16_NACC; j++) s = _mm512_add_ps(s, acc[j]);
-  return (double)_mm512_reduce_add_ps(s);
-}
-#endif
-
-// ---- Mixed precision (fp16 mul -> fp32 acc, widening FMLA) ------------------
-#if (defined(__ARM_FEATURE_FP16FML) || defined(__ARM_FEATURE_FP16_FML)) && defined(__ARM_FEATURE_FP16_VECTOR_ARITHMETIC)
-#define CPU_HAS_MP_KERNEL 1
-static constexpr int MP_NACC = 16, MP_OPS_PER_INSTR = 16;  // 2 FMLAL = 8 mul + 8 add
-static double runMpChain(uint64_t outer)
-{
-  float32x4_t acc[MP_NACC];
-  // FMLAL can only *add* to the fp32 accumulator (its operands are fp16), so a
-  // pure-FMLAL chain `acc += m*b` is a LINEAR function of the iteration count and
-  // -ffast-math closes it into `acc += N*m*b`, deleting every FMLAL -> a peak that
-  // beats the impossible fp16 ceiling (mp must be <= ~0.5x fp16, FMLAL doing half
-  // the flops/instr).  A live-operand recurrence, distinct per-chain coefficients,
-  // and even an asm("+w") barrier all FAILED on the aarch64 server clang -- the
-  // work stays *linear*, so it's always close-form-able.  The robust fix is the
-  // fp16 chain's trick: feed the accumulator back as a MULTIPLICAND so the
-  // recurrence is genuinely NONLINEAR (no closed form exists).  FMLAL multiplies
-  // fp16, so narrow acc -> fp16 each step:
-  //   acc <- acc + narrow(acc)*(-decay) + 1*refill   ->  fixed point refill/decay.
-  // The vcvt is what makes it nonlinear (and uncollapsible); the constant refill
-  // keeps acc off zero so the feedback term stays meaningful.
-  const float16x8_t decay  = vdupq_n_f16((float16_t)-0.000977f);  // <0: contracts
-  const float16x8_t refill = vdupq_n_f16((float16_t) 0.001953f);
-  const float16x8_t one    = vdupq_n_f16((float16_t) 1.0f);
-  for (int j = 0; j < MP_NACC; j++) acc[j] = vdupq_n_f32(1.0f + 0.01f * j);
-  for (uint64_t o = 0; o < outer; o++)
-    CPU_UNROLL_K
-    for (int k = 0; k < INNER; k++)
-    {
-      CPU_UNROLL_FULL
-      for (int j = 0; j < MP_NACC; j++)
-      {
-        float16x4_t h  = vcvt_f16_f32(acc[j]);          // narrow: nonlinear feedback
-        float16x8_t hh = vcombine_f16(h, h);
-        acc[j] = vfmlalq_low_f16(acc[j], hh, decay);    // acc += narrow(acc)*(-decay)
-        acc[j] = vfmlalq_high_f16(acc[j], one, refill); // acc += 1*refill (off-zero)
-      }
-    }
-  float32x4_t s = acc[0];
-  for (int j = 1; j < MP_NACC; j++) s = vaddq_f32(s, acc[j]);
-  return (double)vaddvq_f32(s);
-}
-#endif
-
-// ---- INT8 dot product ------------------------------------------------------
-#if defined(__AVX512VNNI__)
-#define CPU_HAS_INT8DP_KERNEL 1
-static constexpr int I8_NACC = 16, I8_OPS_PER_INSTR = 128;
-static double runInt8DpChain(uint64_t outer)
-{
-  __m512i acc[I8_NACC];
-  const __m512i a = _mm512_set1_epi8((char)3);
-  const __m512i b = _mm512_set1_epi8((char)5);
-  for (int j = 0; j < I8_NACC; j++) acc[j] = _mm512_setzero_si512();
-  for (uint64_t o = 0; o < outer; o++)
-    CPU_UNROLL_K
-    for (int k = 0; k < INNER; k++)
-    {
-      CPU_UNROLL_FULL
-      for (int j = 0; j < I8_NACC; j++) acc[j] = _mm512_dpbusd_epi32(acc[j], a, b);
-    }
-  __m512i s = acc[0];
-  for (int j = 1; j < I8_NACC; j++) s = _mm512_add_epi32(s, acc[j]);
-  return (double)_mm512_reduce_add_epi32(s);
-}
-#elif defined(__AVXVNNI__)
-#define CPU_HAS_INT8DP_KERNEL 1
-static constexpr int I8_NACC = 12, I8_OPS_PER_INSTR = 64;
-static double runInt8DpChain(uint64_t outer)
-{
-  __m256i acc[I8_NACC];
-  const __m256i a = _mm256_set1_epi8((char)3);
-  const __m256i b = _mm256_set1_epi8((char)5);
-  for (int j = 0; j < I8_NACC; j++) acc[j] = _mm256_setzero_si256();
-  for (uint64_t o = 0; o < outer; o++)
-    CPU_UNROLL_K
-    for (int k = 0; k < INNER; k++)
-    {
-      CPU_UNROLL_FULL
-      for (int j = 0; j < I8_NACC; j++) acc[j] = _mm256_dpbusd_epi32(acc[j], a, b);
-    }
-  __m256i sAll = acc[0];
-  for (int j = 1; j < I8_NACC; j++) sAll = _mm256_add_epi32(sAll, acc[j]);
-  __m128i lo = _mm256_castsi256_si128(sAll);
-  __m128i hi = _mm256_extracti128_si256(sAll, 1);
-  lo = _mm_add_epi32(lo, hi);
-  lo = _mm_hadd_epi32(lo, lo);
-  lo = _mm_hadd_epi32(lo, lo);
-  return (double)_mm_cvtsi128_si32(lo);
-}
-#elif defined(__AVXVNNIINT8__)
-// 256-bit AVX-VNNI-INT8: signed×signed int8 dot (VPDPBSSD).  Same throughput as
-// AVX-VNNI's VPDPBUSD but a distinct ISA row (Zen 6, Lunar Lake, Arrow Lake+).
-#define CPU_HAS_INT8DP_KERNEL 1
-static constexpr int I8_NACC = 12, I8_OPS_PER_INSTR = 64;
-static double runInt8DpChain(uint64_t outer)
-{
-  __m256i acc[I8_NACC];
-  const __m256i a = _mm256_set1_epi8((char)3);
-  const __m256i b = _mm256_set1_epi8((char)5);
-  for (int j = 0; j < I8_NACC; j++) acc[j] = _mm256_setzero_si256();
-  for (uint64_t o = 0; o < outer; o++)
-    CPU_UNROLL_K
-    for (int k = 0; k < INNER; k++)
-    {
-      CPU_UNROLL_FULL
-      for (int j = 0; j < I8_NACC; j++) acc[j] = _mm256_dpbssd_epi32(acc[j], a, b);
-    }
-  __m256i sAll = acc[0];
-  for (int j = 1; j < I8_NACC; j++) sAll = _mm256_add_epi32(sAll, acc[j]);
-  __m128i lo = _mm256_castsi256_si128(sAll);
-  __m128i hi = _mm256_extracti128_si256(sAll, 1);
-  lo = _mm_add_epi32(lo, hi);
-  lo = _mm_hadd_epi32(lo, lo);
-  lo = _mm_hadd_epi32(lo, lo);
-  return (double)_mm_cvtsi128_si32(lo);
-}
-#elif defined(__ARM_FEATURE_DOTPROD)
-#define CPU_HAS_INT8DP_KERNEL 1
-static constexpr int I8_NACC = 16, I8_OPS_PER_INSTR = 32;
-static double runInt8DpChain(uint64_t outer)
-{
-  int32x4_t acc[I8_NACC];
-  const int8x16_t a = vdupq_n_s8(3);
-  const int8x16_t b = vdupq_n_s8(5);
-  for (int j = 0; j < I8_NACC; j++) acc[j] = vdupq_n_s32(0);
-  for (uint64_t o = 0; o < outer; o++)
-    CPU_UNROLL_K
-    for (int k = 0; k < INNER; k++)
-    {
-      CPU_UNROLL_FULL
-      for (int j = 0; j < I8_NACC; j++) acc[j] = vdotq_s32(acc[j], a, b);
-    }
-  int32x4_t s = acc[0];
-  for (int j = 1; j < I8_NACC; j++) s = vaddq_s32(s, acc[j]);
-  return (double)vaddvq_s32(s);
-}
-#endif
-
-// ---- Matrix engine: int8 (AMX / SMMLA) and bf16 (AMX / BFMMLA) -------------
-// ops are PER-k; the table multiplies by INNER (the chain loops INNER per outer).
-#if defined(__AMX_INT8__) && (defined(__x86_64__) || defined(_M_X64))
-#define CPU_MAT_INT8_KERNEL 1
-static constexpr double MAT_I8_OPS_PER_K = 4.0 * 16 * 16 * 64 * 2;
-static thread_local bool g_amxI8Cfg = false;
-alignas(64) static int8_t  g_amxA[16 * 64];   // inputs: read-only, shared is fine
-alignas(64) static int8_t  g_amxB[16 * 64];
-// Output is written by every MT worker; thread_local avoids a data race and the
-// cross-thread cache-line invalidation that shared storage would cause.
-alignas(64) static thread_local int32_t g_amxC[16 * 16];
-static void amxConfigTiles()
-{
-  struct { uint8_t palette, start_row, reserved[14]; uint16_t colsb[16]; uint8_t rows[16]; } cfg = {};
-  cfg.palette = 1;
-  for (int t = 0; t < 6; t++) { cfg.rows[t] = 16; cfg.colsb[t] = 64; }
-  _tile_loadconfig(&cfg);
-}
-static double runMatInt8Chain(uint64_t outer)
-{
-  if (!g_amxI8Cfg) { amxConfigTiles(); g_amxI8Cfg = true; }
-  _tile_loadd(4, g_amxA, 64);
-  _tile_loadd(5, g_amxB, 64);
-  _tile_zero(0); _tile_zero(1); _tile_zero(2); _tile_zero(3);
-  for (uint64_t o = 0; o < outer; o++)
-    for (int k = 0; k < INNER; k++)
-    { _tile_dpbssd(0, 4, 5); _tile_dpbssd(1, 4, 5); _tile_dpbssd(2, 4, 5); _tile_dpbssd(3, 4, 5); }
-  double sink = 0.0;
-  _tile_stored(0, g_amxC, 64); sink += g_amxC[0];
-  _tile_stored(1, g_amxC, 64); sink += g_amxC[0];
-  _tile_stored(2, g_amxC, 64); sink += g_amxC[0];
-  _tile_stored(3, g_amxC, 64); sink += g_amxC[0];
-  return sink;
-}
-#elif defined(__ARM_FEATURE_MATMUL_INT8)
-#define CPU_MAT_INT8_KERNEL 1
-static constexpr double MAT_I8_OPS_PER_K = 16.0 * 64.0;   // MM_NACC * 64
-static double runMatInt8Chain(uint64_t outer)
-{
-  constexpr int NACC = 16;
-  int32x4_t acc[NACC];
-  const int8x16_t a = vdupq_n_s8(3);
-  const int8x16_t b = vdupq_n_s8(5);
-  for (int j = 0; j < NACC; j++) acc[j] = vdupq_n_s32(0);
-  for (uint64_t o = 0; o < outer; o++)
-    CPU_UNROLL_K
-    for (int k = 0; k < INNER; k++)
-    {
-      CPU_UNROLL_FULL
-      for (int j = 0; j < NACC; j++) acc[j] = vmmlaq_s32(acc[j], a, b);
-    }
-  int32x4_t s = acc[0];
-  for (int j = 1; j < NACC; j++) s = vaddq_s32(s, acc[j]);
-  return (double)vaddvq_s32(s);
-}
-#endif
-
-#if defined(__AMX_BF16__) && (defined(__x86_64__) || defined(_M_X64))
-#define CPU_MAT_FP_KERNEL 1
-static constexpr double MAT_FP_OPS_PER_K = 4.0 * 16 * 16 * 32 * 2;
-static thread_local bool g_amxBf16Cfg = false;
-alignas(64) static uint16_t g_amxAb[16 * 32];   // inputs: read-only, shared is fine
-alignas(64) static uint16_t g_amxBb[16 * 32];
-// Output is written by every MT worker; thread_local avoids a data race (see int8).
-alignas(64) static thread_local float g_amxCf[16 * 16];
-static void amxConfigTilesBf16()
-{
-  struct { uint8_t palette, start_row, reserved[14]; uint16_t colsb[16]; uint8_t rows[16]; } cfg = {};
-  cfg.palette = 1;
-  for (int t = 0; t < 6; t++) { cfg.rows[t] = 16; cfg.colsb[t] = 64; }
-  _tile_loadconfig(&cfg);
-}
-static double runMatFpChain(uint64_t outer)
-{
-  if (!g_amxBf16Cfg) { amxConfigTilesBf16(); g_amxBf16Cfg = true; }
-  _tile_loadd(4, g_amxAb, 64);
-  _tile_loadd(5, g_amxBb, 64);
-  _tile_zero(0); _tile_zero(1); _tile_zero(2); _tile_zero(3);
-  for (uint64_t o = 0; o < outer; o++)
-    for (int k = 0; k < INNER; k++)
-    { _tile_dpbf16ps(0, 4, 5); _tile_dpbf16ps(1, 4, 5); _tile_dpbf16ps(2, 4, 5); _tile_dpbf16ps(3, 4, 5); }
-  double sink = 0.0;
-  _tile_stored(0, g_amxCf, 64); sink += g_amxCf[0];
-  _tile_stored(1, g_amxCf, 64); sink += g_amxCf[0];
-  _tile_stored(2, g_amxCf, 64); sink += g_amxCf[0];
-  _tile_stored(3, g_amxCf, 64); sink += g_amxCf[0];
-  return sink;
-}
-#elif defined(__ARM_FEATURE_BF16_VECTOR_ARITHMETIC)
-// BFMMLA is part of FEAT_BF16 (no separate I8MM needed).
-#define CPU_MAT_FP_KERNEL 1
-static constexpr double MAT_FP_OPS_PER_K = 16.0 * 32.0;   // BMM_NACC * 32
-static double runMatFpChain(uint64_t outer)
-{
-  constexpr int NACC = 16;
-  float32x4_t acc[NACC];
-  const bfloat16x4_t lo = vcvt_bf16_f32(vdupq_n_f32(1.0001f));
-  const bfloat16x8_t a  = vcombine_bf16(lo, lo);
-  const bfloat16x8_t b  = vcombine_bf16(lo, lo);
-  for (int j = 0; j < NACC; j++) acc[j] = vdupq_n_f32(0.0f);
-  for (uint64_t o = 0; o < outer; o++)
-    CPU_UNROLL_K
-    for (int k = 0; k < INNER; k++)
-    {
-      CPU_UNROLL_FULL
-      for (int j = 0; j < NACC; j++) acc[j] = vbfmmlaq_f32(acc[j], a, b);
-    }
-  float32x4_t s = acc[0];
-  for (int j = 1; j < NACC; j++) s = vaddq_f32(s, acc[j]);
-  return (double)vaddvq_f32(s);
-}
-#endif
-
-// ---- Newer x86 AMX matrix dtypes: fp16 / tf32 / fp8 ------------------------
-// Same 4-accumulator-tile config as the int8/bf16 AMX kernels (opaque tiles, so
-// the accumulate loop can't be scalar-evolved; inputs zero => data-independent
-// throughput; output tiles thread_local so every MT worker gets its own).  The
-// TILECFG is the canonical palette-1 / 16-row / 64-colsb layout for all of them;
-// only the element width (hence K per tile) and the DP intrinsic differ.
-#if (defined(__AMX_FP16__) || defined(__AMX_TF32__) || defined(__AMX_FP8__)) && \
-    (defined(__x86_64__) || defined(_M_X64))
-static void amxConfig16x64()
-{
-  struct { uint8_t palette, start_row, reserved[14]; uint16_t colsb[16]; uint8_t rows[16]; } cfg = {};
-  cfg.palette = 1;
-  for (int t = 0; t < 6; t++) { cfg.rows[t] = 16; cfg.colsb[t] = 64; }
-  _tile_loadconfig(&cfg);
-}
-#endif
-
-#if defined(__AMX_FP16__) && (defined(__x86_64__) || defined(_M_X64))
-#define CPU_MAT_FP16_KERNEL 1
-static constexpr double MAT_FP16_OPS_PER_K = 4.0 * 16 * 16 * 32 * 2;  // K=32 fp16/row
-static thread_local bool g_amxFp16Cfg = false;
-alignas(64) static uint16_t g_amxAf16[16 * 32];
-alignas(64) static uint16_t g_amxBf16m[16 * 32];
-alignas(64) static thread_local float g_amxCf16[16 * 16];
-static double runMatFp16Chain(uint64_t outer)
-{
-  if (!g_amxFp16Cfg) { amxConfig16x64(); g_amxFp16Cfg = true; }
-  _tile_loadd(4, g_amxAf16, 64);
-  _tile_loadd(5, g_amxBf16m, 64);
-  _tile_zero(0); _tile_zero(1); _tile_zero(2); _tile_zero(3);
-  for (uint64_t o = 0; o < outer; o++)
-    for (int k = 0; k < INNER; k++)
-    { _tile_dpfp16ps(0, 4, 5); _tile_dpfp16ps(1, 4, 5); _tile_dpfp16ps(2, 4, 5); _tile_dpfp16ps(3, 4, 5); }
-  double sink = 0.0;
-  _tile_stored(0, g_amxCf16, 64); sink += g_amxCf16[0];
-  _tile_stored(1, g_amxCf16, 64); sink += g_amxCf16[0];
-  _tile_stored(2, g_amxCf16, 64); sink += g_amxCf16[0];
-  _tile_stored(3, g_amxCf16, 64); sink += g_amxCf16[0];
-  return sink;
-}
-#endif
-
-#if defined(__AMX_TF32__) && (defined(__x86_64__) || defined(_M_X64))
-#define CPU_MAT_TF32_KERNEL 1
-static constexpr double MAT_TF32_OPS_PER_K = 4.0 * 16 * 16 * 16 * 2;  // K=16 tf32(fp32-stored)/row
-static thread_local bool g_amxTf32Cfg = false;
-alignas(64) static float g_amxAt32[16 * 16];
-alignas(64) static float g_amxBt32[16 * 16];
-alignas(64) static thread_local float g_amxCt32[16 * 16];
-static double runMatTf32Chain(uint64_t outer)
-{
-  if (!g_amxTf32Cfg) { amxConfig16x64(); g_amxTf32Cfg = true; }
-  _tile_loadd(4, g_amxAt32, 64);
-  _tile_loadd(5, g_amxBt32, 64);
-  _tile_zero(0); _tile_zero(1); _tile_zero(2); _tile_zero(3);
-  for (uint64_t o = 0; o < outer; o++)
-    for (int k = 0; k < INNER; k++)
-    { _tile_mmultf32ps(0, 4, 5); _tile_mmultf32ps(1, 4, 5); _tile_mmultf32ps(2, 4, 5); _tile_mmultf32ps(3, 4, 5); }
-  double sink = 0.0;
-  _tile_stored(0, g_amxCt32, 64); sink += g_amxCt32[0];
-  _tile_stored(1, g_amxCt32, 64); sink += g_amxCt32[0];
-  _tile_stored(2, g_amxCt32, 64); sink += g_amxCt32[0];
-  _tile_stored(3, g_amxCt32, 64); sink += g_amxCt32[0];
-  return sink;
-}
-#endif
-
-#if defined(__AMX_FP8__) && (defined(__x86_64__) || defined(_M_X64))
-#define CPU_MAT_FP8_KERNEL 1
-static constexpr double MAT_FP8_OPS_PER_K = 4.0 * 16 * 16 * 64 * 2;  // K=64 fp8/row
-static thread_local bool g_amxFp8Cfg = false;
-alignas(64) static uint8_t g_amxAf8[16 * 64];
-alignas(64) static uint8_t g_amxBf8[16 * 64];
-alignas(64) static thread_local float g_amxCf8[16 * 16];
-static double runMatFp8Chain(uint64_t outer)
-{
-  if (!g_amxFp8Cfg) { amxConfig16x64(); g_amxFp8Cfg = true; }
-  _tile_loadd(4, g_amxAf8, 64);
-  _tile_loadd(5, g_amxBf8, 64);
-  _tile_zero(0); _tile_zero(1); _tile_zero(2); _tile_zero(3);
-  // hf8 = e4m3 (the common ML fp8); throughput is format-independent.
-  for (uint64_t o = 0; o < outer; o++)
-    for (int k = 0; k < INNER; k++)
-    { _tile_dphf8ps(0, 4, 5); _tile_dphf8ps(1, 4, 5); _tile_dphf8ps(2, 4, 5); _tile_dphf8ps(3, 4, 5); }
-  double sink = 0.0;
-  _tile_stored(0, g_amxCf8, 64); sink += g_amxCf8[0];
-  _tile_stored(1, g_amxCf8, 64); sink += g_amxCf8[0];
-  _tile_stored(2, g_amxCf8, 64); sink += g_amxCf8[0];
-  _tile_stored(3, g_amxCf8, 64); sink += g_amxCf8[0];
-  return sink;
-}
-#endif
-
-// ---- AVX10.2 native bf16 vector FMA (full-rate, not a dot) ------------------
-// Distinct from the AVX512-BF16 dot path: VFMADD*PBF16 does a real bf16 multiply-
-// add at full 512-bit vector rate (32 lanes).  Same affine chain as fp16; b must
-// not round to 1.0 in bf16 (0.98828 is exact and distinct).
-#if defined(__AVX10_2_512__)
-#define CPU_HAS_BF16FMA_KERNEL 1
-static constexpr int BF16FMA_LANES = 32, BF16FMA_NACC = 16;
-static double runBf16FmaChain(uint64_t outer)
-{
-  __m512bh acc[BF16FMA_NACC];
-  const __m512bh b = _mm512_set1_pbh((__bf16)0.98828f);
-  const __m512bh c = _mm512_set1_pbh((__bf16)0.0117f);
-  for (int j = 0; j < BF16FMA_NACC; j++) acc[j] = _mm512_set1_pbh((__bf16)(0.1f * (j + 1)));
-  for (uint64_t o = 0; o < outer; o++)
-    CPU_UNROLL_K
-    for (int k = 0; k < INNER; k++)
-    {
-      CPU_UNROLL_FULL
-      for (int j = 0; j < BF16FMA_NACC; j++) acc[j] = _mm512_fmadd_pbh(acc[j], b, c);
-    }
-  __m512bh s = acc[0];
-  for (int j = 1; j < BF16FMA_NACC; j++) s = _mm512_add_pbh(s, acc[j]);
-  // Reduce via a raw bf16->fp32 widening (no cast intrinsic takes __m512bh whole).
-  uint16_t tmp[32]; std::memcpy(tmp, &s, sizeof(tmp));
-  double r = 0.0;
-  for (int i = 0; i < 32; i++)
-  { uint32_t u = (uint32_t)tmp[i] << 16; float f; std::memcpy(&f, &u, sizeof(f)); r += (double)f; }
-  return r;
-}
-#endif
-
-// ---- ARM SVE (vector-length-agnostic) -------------------------------------
-// SVE registers are "sizeless": svfloat32_t etc. can't be array/struct members,
-// so the NACC independent accumulator chains are declared as individual named
-// registers via the SVE_REP macros instead of the acc[NACC] loop the fixed-width
-// kernels use.  The kernels are VL-agnostic (one binary runs 128-bit Oryon/Vera,
-// 256-bit Graviton3, ...); the op count is therefore VL-dependent and computed
-// from svcntw()/svcntd() at table-build time (only ever reached when SVE is the
-// running host's ISA).  b/c coefficients ride in loop-invariant Z registers, so
-// each svmad is one back-to-back MAD with no reload -- the SVE analogue of the
-// NEON FMLA chain.
-#if defined(__ARM_FEATURE_SVE) && !defined(CLPEAK_CORE_ONLY)
-#define CPU_HAS_SVE_KERNELS 1
-
-// Expand X(i) for i in [0, N).  32 architectural Z registers give headroom for
-// 24 fp accumulators + the 2 coefficient regs (same latency-hiding rationale as
-// the NEON NACC=24; revisit with a sweep when validating on real SVE silicon).
-#define SVE_REP16(X) X(0)X(1)X(2)X(3)X(4)X(5)X(6)X(7)X(8)X(9)X(10)X(11)X(12)X(13)X(14)X(15)
-#define SVE_REP24(X) SVE_REP16(X) X(16)X(17)X(18)X(19)X(20)X(21)X(22)X(23)
-static constexpr int SVE_NACC_FP  = 24;   // fp32/fp64/int32 MAD chains
-static constexpr int SVE_NACC_DOT = 16;   // SDOT / BFDOT / *MMLA (wider per-instr)
-
-// acc = acc*b + c  (svmad: MAD Zdn = Zdn*Zm + Za, destructive on the accumulator)
-static double runSveFp32Chain(uint64_t outer)
-{
-  const svbool_t pg = svptrue_b32();
-  volatile float vb = 0.999999f, vc = 0.000001f;
-  const svfloat32_t b = svdup_f32(vb), c = svdup_f32(vc);
-#define DECL(i) svfloat32_t a##i = svdup_f32(0.1f * (float)((i) + 1));
-  SVE_REP24(DECL)
-#undef DECL
-  for (uint64_t o = 0; o < outer; o++)
-    CPU_UNROLL_K
-    for (int k = 0; k < INNER; k++)
-    {
-#define STEP(i) a##i = svmad_f32_x(pg, a##i, b, c);
-      SVE_REP24(STEP)
-#undef STEP
-    }
-  svfloat32_t s = svdup_f32(0.0f);
-#define RED(i) s = svadd_f32_x(pg, s, a##i);
-  SVE_REP24(RED)
-#undef RED
-  return (double)svaddv_f32(pg, s);
-}
-
-static double runSveFp64Chain(uint64_t outer)
-{
-  const svbool_t pg = svptrue_b64();
-  volatile double vb = 0.999999, vc = 0.000001;
-  const svfloat64_t b = svdup_f64(vb), c = svdup_f64(vc);
-#define DECL(i) svfloat64_t a##i = svdup_f64(0.1 * (double)((i) + 1));
-  SVE_REP24(DECL)
-#undef DECL
-  for (uint64_t o = 0; o < outer; o++)
-    CPU_UNROLL_K
-    for (int k = 0; k < INNER; k++)
-    {
-#define STEP(i) a##i = svmad_f64_x(pg, a##i, b, c);
-      SVE_REP24(STEP)
-#undef STEP
-    }
-  svfloat64_t s = svdup_f64(0.0);
-#define RED(i) s = svadd_f64_x(pg, s, a##i);
-  SVE_REP24(RED)
-#undef RED
-  return (double)svaddv_f64(pg, s);
-}
-
-static double runSveInt32Chain(uint64_t outer)
-{
-  const svbool_t pg = svptrue_b32();
-  volatile int vmul = 1664525;
-  const svint32_t b = svdup_s32(vmul), c = svdup_s32(1013904223);
-#define DECL(i) svint32_t a##i = svdup_s32((i) + 1);
-  SVE_REP24(DECL)
-#undef DECL
-  for (uint64_t o = 0; o < outer; o++)
-    CPU_UNROLL_K
-    for (int k = 0; k < INNER; k++)
-    {
-#define STEP(i) a##i = svmad_s32_x(pg, a##i, b, c);
-      SVE_REP24(STEP)
-#undef STEP
-    }
-  svint32_t s = svdup_s32(0);
-#define RED(i) s = svadd_s32_x(pg, s, a##i);
-  SVE_REP24(RED)
-#undef RED
-  return (double)svaddv_s32(pg, s);
-}
-
-// int8 4-way dot product into int32 lanes (SDOT: base SVE, no SVE2 needed).
-static double runSveInt8DpChain(uint64_t outer)
-{
-  const svbool_t pg = svptrue_b32();
-  const svint8_t a = svdup_s8(3), b = svdup_s8(5);
-#define DECL(i) svint32_t acc##i = svdup_s32(0);
-  SVE_REP16(DECL)
-#undef DECL
-  for (uint64_t o = 0; o < outer; o++)
-    CPU_UNROLL_K
-    for (int k = 0; k < INNER; k++)
-    {
-#define STEP(i) acc##i = svdot_s32(acc##i, a, b);
-      SVE_REP16(STEP)
-#undef STEP
-    }
-  svint32_t s = svdup_s32(0);
-#define RED(i) s = svadd_s32_x(pg, s, acc##i);
-  SVE_REP16(RED)
-#undef RED
-  return (double)svaddv_s32(pg, s);
-}
-#endif // __ARM_FEATURE_SVE
-
-// SVE BF16: BFDOT (2-way bf16 dot -> fp32) and BFMMLA (2x2 matrix accumulate).
-#if defined(__ARM_FEATURE_SVE_BF16) && !defined(CLPEAK_CORE_ONLY)
-#define CPU_HAS_SVE_BF16_KERNEL 1
-static double runSveBf16Chain(uint64_t outer)
-{
-  const svbool_t pg = svptrue_b32();
-  // 0x3F81 ~= 1.0039 in bf16 (nonzero, distinct from an exact 1.0 collapse).
-  const svbfloat16_t a = svreinterpret_bf16_u16(svdup_u16(0x3F81));
-  const svbfloat16_t b = a;
-#define DECL(i) svfloat32_t acc##i = svdup_f32(0.0f);
-  SVE_REP16(DECL)
-#undef DECL
-  for (uint64_t o = 0; o < outer; o++)
-    CPU_UNROLL_K
-    for (int k = 0; k < INNER; k++)
-    {
-#define STEP(i) acc##i = svbfdot_f32(acc##i, a, b);
-      SVE_REP16(STEP)
-#undef STEP
-    }
-  svfloat32_t s = svdup_f32(0.0f);
-#define RED(i) s = svadd_f32_x(pg, s, acc##i);
-  SVE_REP16(RED)
-#undef RED
-  return (double)svaddv_f32(pg, s);
-}
-static double runSveMatBf16Chain(uint64_t outer)
-{
-  const svbool_t pg = svptrue_b32();
-  const svbfloat16_t a = svreinterpret_bf16_u16(svdup_u16(0x3F81));
-  const svbfloat16_t b = a;
-#define DECL(i) svfloat32_t acc##i = svdup_f32(0.0f);
-  SVE_REP16(DECL)
-#undef DECL
-  for (uint64_t o = 0; o < outer; o++)
-    CPU_UNROLL_K
-    for (int k = 0; k < INNER; k++)
-    {
-#define STEP(i) acc##i = svbfmmla_f32(acc##i, a, b);
-      SVE_REP16(STEP)
-#undef STEP
-    }
-  svfloat32_t s = svdup_f32(0.0f);
-#define RED(i) s = svadd_f32_x(pg, s, acc##i);
-  SVE_REP16(RED)
-#undef RED
-  return (double)svaddv_f32(pg, s);
-}
-#endif // __ARM_FEATURE_SVE_BF16
-
-// SVE I8MM: SMMLA (int8 2x2 matrix accumulate into int32).
-#if defined(__ARM_FEATURE_SVE_MATMUL_INT8) && !defined(CLPEAK_CORE_ONLY)
-#define CPU_HAS_SVE_I8MM_KERNEL 1
-static double runSveMatInt8Chain(uint64_t outer)
-{
-  const svbool_t pg = svptrue_b32();
-  const svint8_t a = svdup_s8(3), b = svdup_s8(5);
-#define DECL(i) svint32_t acc##i = svdup_s32(0);
-  SVE_REP16(DECL)
-#undef DECL
-  for (uint64_t o = 0; o < outer; o++)
-    CPU_UNROLL_K
-    for (int k = 0; k < INNER; k++)
-    {
-#define STEP(i) acc##i = svmmla_s32(acc##i, a, b);
-      SVE_REP16(STEP)
-#undef STEP
-    }
-  svint32_t s = svdup_s32(0);
-#define RED(i) s = svadd_s32_x(pg, s, acc##i);
-  SVE_REP16(RED)
-#undef RED
-  return (double)svaddv_s32(pg, s);
-}
-#endif // __ARM_FEATURE_SVE_MATMUL_INT8
-
-#endif // CLPEAK_CORE_ONLY
 
 // ---- Table builder (this TU's offered kernels) ----------------------------
 static const CpuKernelTable *tuTable()
